MDAssignment/requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
Django==3.2.16
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
google==3.0.0
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.13.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.4
importlib-metadata==5.0.0
Markdown==3.4.1
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.6.0
pyparsing==3.0.9
pytz==2022.5
requests==2.28.1
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
typing-extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
zipp==3.10.0
